Capitulo 7 - Normalizando o tempo
Aplica normalização de tempo para que o movimento ocorra de maneira fixa e o número de frames fique fixo.

+++ src/Bola.java
@@ -14,11 +14,11 @@ public class Bola {
     public BufferedImage direita_cima;
     public BufferedImage esquerda_baixo;
     public BufferedImage esquerda_cima;
-    public int posX;
-    public int posY;
-    public int raio;
-    public int velX;
-    public int velY;
+    public double posX;
+    public double posY;
+    public double raio;
+    public double velX;
+    public double velY;
 
     public Bola() {
         try {
@@ -72,11 +72,11 @@ public class Bola {
         return image;
     }
 
-    public int getCentroX() {
+    public double getCentroX() {
         return posX + raio;
     }
 
-    public int getCentroY() {
+    public double getCentroY() {
         return posY + raio;
     }
 }
